Added async-await

+++ server/server.js
@@ -58,25 +58,46 @@
         })
     });
 
-    app.delete('/todos/:id',authenticate,(req,res) => {
-       var id = req.params.id;
-       //Validate _id
-       if (!ObjectId.isValid(id)) {
-           return res.status(404).send();
-       }
-       //Find by the id and remove
-       Todo.findOneAndRemove({
-           _id:id,
-           _creator:req.user._id
-        }).then((todo) => {
-        //Validate if a todo was found
+    app.delete('/todos/:id',authenticate, async (req,res) => {
+
+        try {
+            var id = req.params.id;
+            //Validate _id
+            if (!ObjectId.isValid(id)) {
+                return res.status(404).send();
+            }
+            //Find by the id and remove
+
+            const todo =  await Todo.findOneAndRemove({_id: id,_creator: req.user._id});
+
             if (!todo) {
                 return res.status(404).send();
             }
             res.send({todo});
-       },(e) => {
-            return res.status(400).send(e);
-       })
+        } catch (e) {
+            res.status(400).send(e);
+        }
+    //    var id = req.params.id;
+    //    //Validate _id
+    //    if (!ObjectId.isValid(id)) {
+    //        return res.status(404).send();
+    //    }
+    //    //Find by the id and remove
+
+       
+
+    //    Todo.findOneAndRemove({
+    //        _id:id,
+    //        _creator:req.user._id
+    //     }).then((todo) => {
+    //     //Validate if a todo was found
+    //         if (!todo) {
+    //             return res.status(404).send();
+    //         }
+    //         res.send({todo});
+    //    },(e) => {
+    //         return res.status(400).send(e);
+    //    })
     })
 
     app.patch('/todos/:id',authenticate,(req,res) => {
@@ -109,20 +130,28 @@
         })
     })
 //USERS
-    app.post('/users',(req,res) => {
+    app.post('/users', async (req,res) => {
 
-        var body = _.pick(req.body, ['email','password']);
-        
-        var user = new User(body);
+        try {
+            const body = _.pick(req.body, ['email','password']);
+            const user = new User(body);
+            await user.save();
+            const token = user.generateAuthToken();
+            res.header('x-auth', token).send(user);
+        } catch (e) {
+            res.status(400).send(e);
+        }
 
-        user.save().then(() => {
-            return user.generateAuthToken();
-        }).then((token) => {
-            res.header('x-auth',token).send(user);
-        })
-        .catch((e) =>  {
-            res.status(400).send(e)
-        });
+
+
+        // user.save().then(() => {
+        //     return user.generateAuthToken();
+        // }).then((token) => {
+        //     res.header('x-auth',token).send(user);
+        // })
+        // .catch((e) =>  {
+        //     res.status(400).send(e)
+        // });
     });
 
    
@@ -131,24 +160,41 @@
            res.send(req.user);
     });
 
-    app.post('/users/login',(req,res) => {
-        let body = _.pick(req.body,['email','password']);
+    app.post('/users/login',async(req,res) => {
+        // const body = _.pick(req.body,['email','password']),
 
-        User.findByCredentials(body.email,body.password).then((user) => {
-           return user.generateAuthToken().then((token) => {
-                res.header('x-auth', token).send(user);
-           })
-        }).catch((e) => {
-            res.status(400).send(e);
-            console.log(e);
+
+             try {
+                const body = _.pick(req.body,['email','password']),
+                user = await User.findByCredentials(body.email, body.password),
+                token = await user.generateAuthToken();
+                res.header('x-auth', token).send(user); 
+
+             } catch (error) {
+                res.status(400).send(error);
+                console.log(error);
+             }
+
+        // User.findByCredentials(body.email,body.password).then((user) => {
+        //    return user.generateAuthToken().then((token) => {
+        //         res.header('x-auth', token).send(user);
+        //    })
+        // }).catch((e) => {
+        //     res.status(400).send(e);
+        //     console.log(e);
             
-        })
+        // })
     });
 
-    app.delete('/users/me/token', authenticate ,(req,res) => {
-       req.user.removeToken(req.token).then(() => {
-           res.status(200).send()
-       }).catch((e) => res.status(400).send())
+    app.delete('/users/me/token', authenticate , async (req,res) => {
+
+        try {
+           await req.user.removeToken(req.token);
+           res.status(200).send();
+        } catch (error) {
+            res.status(400).send()
+        }
+
     });
 
     app.listen(port,() => {
